requirement.txt
Flask==3.0.0
gunicorn==21.2.0
tensorflow-cpu==2.15.0
numpy==1.24.3
scikit-learn==1.3.2
h5py==3.10.0